🤖 Update SDK to version v2.37.4

+++ src/main/java/ai/reveng/model/Symbols.java
@@ -21,6 +21,7 @@ import com.google.gson.annotations.SerializedName;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
 import java.io.IOException;
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -57,7 +58,7 @@ public class Symbols {
   public static final String SERIALIZED_NAME_BASE_ADDRESS = "base_address";
   @SerializedName(SERIALIZED_NAME_BASE_ADDRESS)
   @javax.annotation.Nonnull
-  private Long baseAddress;
+  private BigInteger baseAddress;
 
   public static final String SERIALIZED_NAME_FUNCTION_BOUNDARIES = "function_boundaries";
   @SerializedName(SERIALIZED_NAME_FUNCTION_BOUNDARIES)
@@ -67,7 +68,7 @@ public class Symbols {
   public Symbols() {
   }
 
-  public Symbols baseAddress(@javax.annotation.Nonnull Long baseAddress) {
+  public Symbols baseAddress(@javax.annotation.Nonnull BigInteger baseAddress) {
     this.baseAddress = baseAddress;
     return this;
   }
@@ -77,11 +78,11 @@ public class Symbols {
    * @return baseAddress
    */
   @javax.annotation.Nonnull
-  public Long getBaseAddress() {
+  public BigInteger getBaseAddress() {
     return baseAddress;
   }
 
-  public void setBaseAddress(@javax.annotation.Nonnull Long baseAddress) {
+  public void setBaseAddress(@javax.annotation.Nonnull BigInteger baseAddress) {
     this.baseAddress = baseAddress;
   }
